requirements.txt
chainlit
openai     
python-dotenv   
psycopg2-binary   
sqlalchemy   
semantic-kernel
asyncpg